requirements.txt
requests
feedparser
google-generativeai
schedule
python-dotenv
beautifulsoup4
python-dateutil
yfinance